+++ datasources/parquet/reconstruct.go
@@ -293,9 +293,9 @@ func assignValue(dst *octosql.Value, src parquet.Value) error {
 	case parquet.Boolean:
 		*dst = octosql.NewBoolean(src.Boolean())
 	case parquet.Int32:
-		*dst = octosql.NewInt(int(src.Int32()))
+		*dst = octosql.NewInt(src.Int64())
 	case parquet.Int64:
-		*dst = octosql.NewInt(int(src.Int64()))
+		*dst = octosql.NewInt(src.Int64())
 	case parquet.Int96:
 		*dst = octosql.NewString(src.Int96().String())
 	case parquet.Float:
